v0.1.0
Rewrote to be more generic.

The previous version required changes and compile switches to work for
different chips. The new code can be used on any PIC chip with no
changes.

Changed the revision control structure. The project config file now
defines the version of library to use which is then checked locally in
the library file. The config file however is required to check for the
presence of the library.